# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Open4Printing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotes:
    post:
      operationId: createQuoteRequest
      tags: [quotes]
      summary: Submit a quote request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteBody"
      responses:
        "201":
          description: Quote request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/quotes:
    get:
      operationId: listAdminQuotes
      tags: [admin]
      summary: List quote requests
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Quote requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuoteRequest"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/quotes/{id}:
    get:
      operationId: getAdminQuote
      tags: [admin]
      summary: Get a single quote request
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Quote request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequestDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAdminQuote
      tags: [admin]
      summary: Update quote status, admin notes, quoted amount
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuoteBody"
      responses:
        "200":
          description: Updated quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequestDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    UploadedFileInfo:
      type: object
      properties:
        id: { type: integer }
        originalName: { type: string }
        fileType: { type: string }
        fileSize: { type: integer }
        side:
          type: ["string", "null"]
          enum: [front, back, null]
        isProof: { type: boolean }
        uploadedAt: { type: string, format: date-time }
      required: [id, originalName, fileType, fileSize, uploadedAt]
...
    CreateQuoteBody:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        productCategory: { type: string }
        productSlug: { type: ["string", "null"] }
        description: { type: string }
        requestedQuantity: { type: ["integer", "null"] }
        notes: { type: string }
        uploadedFileIds:
          type: array
          items: { type: integer }
      required: [name, email, productCategory, description]
    QuoteRequest:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        productCategory: { type: string }
        productSlug: { type: ["string", "null"] }
        description: { type: string }
        requestedQuantity: { type: ["integer", "null"] }
        notes: { type: string }
        status: { type: string }
        adminNotes: { type: string }
        quotedAmount: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, email, productCategory, description, notes, status, adminNotes, createdAt, updatedAt]
    QuoteRequestDetail:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        productCategory: { type: string }
        productSlug: { type: ["string", "null"] }
        description: { type: string }
        requestedQuantity: { type: ["integer", "null"] }
        notes: { type: string }
        status: { type: string }
        adminNotes: { type: string }
        quotedAmount: { type: ["string", "null"] }
        files:
          type: array
          items: { $ref: "#/components/schemas/UploadedFileInfo" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, email, productCategory, description, notes, status, adminNotes, files, createdAt, updatedAt]
    UpdateQuoteBody:
      type: object
      properties:
        status: { type: ["string", "null"] }
        adminNotes: { type: ["string", "null"] }
        quotedAmount: { type: ["string", "null"] }
...