# lib/api-zod/src/generated/types/quoteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Open4Printing API
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteRequest {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  productCategory: string;
  /** @nullable */
  productSlug?: string | null;
  description: string;
  /** @nullable */
  requestedQuantity?: number | null;
  notes: string;
  status: string;
  adminNotes: string;
  /** @nullable */
  quotedAmount?: string | null;
  createdAt: Date;
  updatedAt: Date;
}