# lib/api-zod/src/generated/types/updateQuoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Open4Printing API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateQuoteBody {
  /** @nullable */
  status?: string | null;
  /** @nullable */
  adminNotes?: string | null;
  /** @nullable */
  quotedAmount?: string | null;
}